+++ backend/environment/tests/tests_temp_trend_list.py
@@ -1,0 +1,1 @@
+from django.test import TestCase
